fix: player potion count type mistake

+++ Unity/Entity/Element/Component/Player/Worker/Player Stats/Player OnChain Stats/PlayerOnChainStats.cs
@@ -11,7 +11,7 @@ namespace ForgottenEmpires.Entities.Elements.PlayerWorkers
         public PlayerStats(PlayerStats playerStats) => this.playerStats = playerStats;
 
         // Update onchain data
-        public void UpdateOnChainData(float potionCount)
+        public void UpdateOnChainData(int potionCount)
         {
             this.potionCount = potionCount;
         }
